requirements.txt
dj-database-url==0.4.2
Django==1.11.5
django-flat-responsive==2.0
django-rest-framework==0.1.0
djangorestframework==3.6.4
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pytz==2017.2
whitenoise==3.3.1
